Refactored duplicated function
